# apps/api/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: GeoPulse API
  version: 1.0.0
  description: |
    GeoPulse backend — orchestrates IP intelligence, crypto market data, and crypto news
    behind a standard `{ data, meta, error }` envelope.
servers:
  - url: /
    description: Current host
...
paths:
...
  /api/market:
    get:
      tags: [Market]
      summary: Market coins
      operationId: getMarket
      parameters:
        - $ref: '#/components/parameters/VsCurrency'
        - name: limit
          in: query
          required: false
          description: Number of coins (1..MARKET_MAX_LIMIT). Default MARKET_DEFAULT_LIMIT (20).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Coin[] envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoinListResponse'
              example:
                data:
                  - id: bitcoin
                    symbol: btc
                    name: Bitcoin
                    image: 'https://example.com/btc.png'
                    currentPrice: 60000
                    currency: usd
                    marketCap: 1000000000
                    marketCapRank: 1
                    priceChangePct24h: 2.5
                    totalVolume: 50000000
                    high24h: 61000
                    low24h: 59000
                    lastUpdated: '2026-07-16T12:00:00.000Z'
                meta:
                  requestId: b1f0c2a4-0000-4000-8000-000000000002
                  source: live
                  provider: coingecko
                  latencyMs: 120
                  cached: false
                error: null
        '400':
          $ref: '#/components/responses/ValidationError'
        '502':
          $ref: '#/components/responses/UpstreamError'
  /api/trending:
    get:
      tags: [Market]
      summary: Trending coins with gainers and losers
      operationId: getTrending
      parameters:
        - $ref: '#/components/parameters/VsCurrency'
      responses:
        '200':
          description: TrendingResult envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendingResultResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '502':
          $ref: '#/components/responses/UpstreamError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Authentication ID token
  parameters:
...
    VsCurrency:
      name: vs
      in: query
      required: false
      description: ISO 4217 quote currency (lowercased)
      schema:
        type: string
        pattern: '^[A-Za-z]{3}$'
        default: usd
        example: usd
  responses:
    ValidationError:
      description: VALIDATION_ERROR
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              requestId: b1f0c2a4-0000-4000-8000-000000000099
              source: live
              latencyMs: 3
              cached: false
            error:
              code: VALIDATION_ERROR
              message: 'Invalid query parameter'
              details:
                - path: vs
                  issue: must be a 3-letter code
...
    UpstreamError:
      description: UPSTREAM_ERROR
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - FORBIDDEN
        - NOT_FOUND
        - RATE_LIMITED
        - UPSTREAM_TIMEOUT
        - UPSTREAM_ERROR
        - CIRCUIT_OPEN
        - INTERNAL
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details: {}
    ResponseMeta:
      type: object
      required: [requestId, source, latencyMs, cached]
      properties:
        requestId:
          type: string
          format: uuid
        source:
          type: string
          enum: [live, cache-l1, cache-l2, fallback]
        provider:
          type: string
        latencyMs:
          type: number
        cached:
          type: boolean
        confidence:
          type: ['number', 'null']
        lastUpdated:
          type: ['string', 'null']
          format: date-time
        degraded:
          type: boolean
    ErrorEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          type: 'null'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        error:
          $ref: '#/components/schemas/ApiError'
...
    Coin:
      type: object
      required:
        - id
        - symbol
        - name
        - image
        - currentPrice
        - currency
        - marketCap
        - marketCapRank
        - priceChangePct24h
        - totalVolume
        - high24h
        - low24h
        - lastUpdated
      properties:
        id:
          type: string
        symbol:
          type: string
        name:
          type: string
        image:
          type: ['string', 'null']
        currentPrice:
          type: number
        currency:
          type: string
        marketCap:
          type: ['number', 'null']
        marketCapRank:
          type: ['number', 'null']
        priceChangePct24h:
          type: ['number', 'null']
        totalVolume:
          type: ['number', 'null']
        high24h:
          type: ['number', 'null']
        low24h:
          type: ['number', 'null']
        lastUpdated:
          type: string
          format: date-time
    TrendingCoin:
      type: object
      required: [id, name, symbol, marketCapRank, thumb]
      properties:
        id:
          type: string
        name:
          type: string
        symbol:
          type: string
        marketCapRank:
          type: ['number', 'null']
        thumb:
          type: ['string', 'null']
    TrendingResult:
      type: object
      required: [trending, gainers, losers]
      properties:
        trending:
          type: array
          items:
            $ref: '#/components/schemas/TrendingCoin'
        gainers:
          type: array
          items:
            $ref: '#/components/schemas/Coin'
        losers:
          type: array
          items:
            $ref: '#/components/schemas/Coin'
...
    CoinListResponse:
      type: object
      required: [data, meta, error]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Coin'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        error:
          type: 'null'
    TrendingResultResponse:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/TrendingResult'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        error:
          type: 'null'
...